+++ src/LionWeb.Generator/Impl/FeatureGeneratorReference.cs
@@ -25,38 +25,86 @@ using Core.Notification.Partition.Emitter;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Names;
-using System.Collections.Immutable;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 using static AstExtensions;
 
 public class FeatureGeneratorReference(Classifier classifier, Reference reference, INames names, LionWebVersions lionWebVersion, GeneratorConfig config) : FeatureGeneratorLinkBase(classifier, reference, names, lionWebVersion, config)
 {
-    public IEnumerable<MemberDeclarationSyntax> RequiredSingleReference() =>
-        new List<MemberDeclarationSyntax>
+    public IEnumerable<MemberDeclarationSyntax> RequiredSingleReference()
+    {
+        TypeSyntax returnType;
+        ExpressionSyntax getter;
+        ExpressionSyntax setter;
+        ExpressionSyntax tryGetReturn;
+        switch (config.UnresolvedReferenceHandling)
+        {
+            case UnresolvedReferenceHandling.Throw:
+                returnType = AsType(reference.GetFeatureType());
+                getter = NotNullOrThrow(ReferenceTargetNonNullTargetCall(),
+                    NewCall([MetaProperty(reference)], AsType(typeof(UnsetFeatureException)))
+                );
+                setter = InvocationExpression(FeatureSet(), AsArguments([IdentifierName("value")]));
+                tryGetReturn = SingleTryGetParamNotNull();
+                break;
+
+            case UnresolvedReferenceHandling.ReturnAsNull:
+                returnType = NullableType(AsType(reference.GetFeatureType()));
+                getter = CallGeneric("GetRequiredReference", AsType(reference.GetFeatureType()),
+                    FeatureField(reference), MetaProperty(reference));
+                setter = InvocationExpression(FeatureSet(), AsArguments([
+                    NotNullOrThrow(IdentifierName("value"),
+                        NewCall([MetaProperty(reference), IdentifierName("value")],
+                            AsType(typeof(InvalidValueException))))
+                ]));
+                tryGetReturn = SingleFieldOrTryGetParamNotNull();
+                break;
+
+            default:
+                throw new ArgumentOutOfRangeException(nameof(config.UnresolvedReferenceHandling),
+                    config.UnresolvedReferenceHandling.ToString());
+        }
+
+        return new List<MemberDeclarationSyntax>
             {
                 SingleReferenceField(),
-                SingleRequiredFeatureProperty(InvocationExpression(ReferenceTarget()))
+                SingleRequiredFeatureProperty(returnType, getter, setter)
                     .Xdoc(XdocThrowsIfSetToNull()),
-                SingleReferenceTargetMethod(),
-                TryGet(InvocationExpression(ReferenceTarget())),
+                TryGet(ReferenceTargetNullableTargetCall(), tryGetReturn),
                 SingleReferenceSetter([
-                        ExpressionStatement(CallGeneric("AssureNotNullInstance",
-                            AsType(reference.GetFeatureType()),
-                            IdentifierName("value"),
-                            MetaProperty(reference)
-                        )),
-                        ReferenceEmitterVariable(),
-                        EmitterCollectOldDataCall(),
-                        AssignFeatureField(),
-                        EmitterNotifyCall(),
-                        ReturnStatement(This())
-                    ])
+                    ExpressionStatement(CallGeneric("AssureNotNullInstance",
+                        AsType(reference.GetFeatureType()),
+                        IdentifierName("value"),
+                        MetaProperty(reference)
+                    )),
+                    ReferenceEmitterVariable(),
+                    EmitterCollectOldDataCall(),
+                    AssignFeatureField(),
+                    EmitterNotifyCall(),
+                    ReturnStatement(This())
+                ])
             }
             .Concat(RequiredFeatureSetter([
                     SingleReferenceSetterForwarder()
                 ])
                 .Select(s => s.Xdoc(XdocThrowsIfSetToNull()))
             );
+    }
+
+    private BinaryExpressionSyntax SingleFieldOrTryGetParamNotNull() =>
+        Or(
+            NotEquals(FeatureField(reference), Null()),
+            SingleTryGetParamNotNull()
+        );
+
+    private BinaryExpressionSyntax SingleTryGetParamNotNull() => NotEquals(IdentifierName(FeatureTryGetParam()), Null());
+
+    private InvocationExpressionSyntax ReferenceTargetNullableTargetCall() =>
+        CallGeneric("ReferenceTargetNullableTarget",
+            AsType(reference.Type), FeatureField(reference), MetaProperty(reference));
+
+    private InvocationExpressionSyntax ReferenceTargetNonNullTargetCall() =>
+        CallGeneric("ReferenceTargetNonNullTarget",
+            AsType(reference.Type), FeatureField(reference), MetaProperty(reference));
 
     private ReturnStatementSyntax SingleReferenceSetterForwarder() =>
         ReturnStatement(Call(FeatureSet().ToString(), FromNodeOptional(IdentifierName("value")),
@@ -78,13 +126,31 @@ public class FeatureGeneratorReference(Classifier classifier, Reference referenc
             ])
         );
 
-    public IEnumerable<MemberDeclarationSyntax> OptionalSingleReference() =>
-        new List<MemberDeclarationSyntax>
+    public IEnumerable<MemberDeclarationSyntax> OptionalSingleReference()
+    {
+        ExpressionSyntax getter;
+        ExpressionSyntax tryGetReturn;
+        switch (config.UnresolvedReferenceHandling)
+        {
+            case UnresolvedReferenceHandling.Throw:
+                getter = ReferenceTargetNonNullTargetCall();
+                tryGetReturn = SingleTryGetParamNotNull();
+                break;
+
+            case UnresolvedReferenceHandling.ReturnAsNull:
+                getter = ReferenceTargetNullableTargetCall();
+                tryGetReturn = SingleFieldOrTryGetParamNotNull();
+                break;
+
+            default:
+                throw new ArgumentOutOfRangeException(config.UnresolvedReferenceHandling.ToString());
+        }
+
+        return new List<MemberDeclarationSyntax>
             {
                 SingleReferenceField(),
-                SingleOptionalFeatureProperty(InvocationExpression(ReferenceTarget())),
-                SingleReferenceTargetMethod(),
-                TryGet(InvocationExpression(ReferenceTarget())),
+                SingleOptionalFeatureProperty(getter),
+                TryGet(ReferenceTargetNullableTargetCall(), tryGetReturn),
                 SingleReferenceSetter([
                     ExpressionStatement(CallGeneric("AssureNullableInstance",
                         AsType(reference.GetFeatureType()),
@@ -101,6 +167,7 @@ public class FeatureGeneratorReference(Classifier classifier, Reference referenc
             .Concat(OptionalFeatureSetter([
                 SingleReferenceSetterForwarder()
             ]));
+    }
 
     private MethodDeclarationSyntax SingleReferenceSetter(List<StatementSyntax> body) =>
         Method(FeatureSet().ToString(), AsType(classifier),
@@ -111,9 +178,6 @@ public class FeatureGeneratorReference(Classifier classifier, Reference referenc
             )
             .WithModifiers(AsModifiers(SyntaxKind.PrivateKeyword))
             .WithBody(AsStatements(body));
-
-    private static InvocationExpressionSyntax FromNodeOptionalValue() =>
-        FromNodeOptional(IdentifierName("value"));
 
     private static InvocationExpressionSyntax FromNodeOptional(ExpressionSyntax value) =>
         InvocationExpression(
@@ -127,17 +191,37 @@ public class FeatureGeneratorReference(Classifier classifier, Reference referenc
                 ])
             );
 
-    public IEnumerable<MemberDeclarationSyntax> RequiredMultiReference() =>
-        new List<MemberDeclarationSyntax>
+    public IEnumerable<MemberDeclarationSyntax> RequiredMultiReference()
+    {
+        string getterMethod;
+        MethodDeclarationSyntax tryGet;
+        switch (config.UnresolvedReferenceHandling)
+        {
+            case UnresolvedReferenceHandling.ReturnAsNull:
+                getterMethod = "GetRequiredNullableReferences";
+                tryGet = TryGetMultiple(
+                    ReferenceTargetNullableTargetsCall(),
+                    NullableType(AsType(reference.GetFeatureType()))
+                );
+                break;
+
+            case UnresolvedReferenceHandling.Throw:
+                getterMethod = "GetRequiredNonNullReferences";
+                tryGet = TryGetMultipleReferences();
+                break;
+
+            default:
+                throw new ArgumentOutOfRangeException(nameof(config.UnresolvedReferenceHandling),
+                    config.UnresolvedReferenceHandling.ToString());
+        }
+
+        return new List<MemberDeclarationSyntax>
         {
             MultipleReferenceField(),
-            MultipleReferenceProperty(CallGeneric("AsNonEmptyReadOnly", AsType(reference.Type),
-                        Call(ReferenceTargets().ToString()),
-                        MetaProperty(reference)
-                    ))
+            MultipleReferenceProperty(CallGeneric(getterMethod, AsType(reference.GetFeatureType()),
+                    FeatureField(reference), MetaProperty(reference)))
                 .Xdoc(XdocRequiredMultipleLink()),
-            MultipleReferenceTargetsMethod(),
-            TryGetMultiple(InvocationExpression(ReferenceTargets()))
+            tryGet
         }.Concat(
             LinkAdder([
                     SafeNodesVariableReference(),
@@ -171,7 +255,7 @@ public class FeatureGeneratorReference(Classifier classifier, Reference referenc
                     AssureNonEmptyCall(),
                     ExpressionStatement(Call("AssureNotClearing",
                         IdentifierName("safeNodes"),
-                        InvocationExpression(ReferenceTargets()),
+                        ReferenceTargetNullableTargetsCall(),
                         MetaProperty(reference)
                     )),
                     SimpleRemoveAllCall(),
@@ -179,6 +263,7 @@ public class FeatureGeneratorReference(Classifier classifier, Reference referenc
                 ])
                 .Select(XdocRequiredRemover)
         );
+    }
 
     private LocalDeclarationStatementSyntax AddMultipleReferenceEmitterVariable(ExpressionSyntax index) =>
         Variable(
@@ -190,46 +275,81 @@ public class FeatureGeneratorReference(Classifier classifier, Reference referenc
             ])
         );
 
-    public IEnumerable<MemberDeclarationSyntax> OptionalMultiReference() =>
-        new List<MemberDeclarationSyntax>
+    public IEnumerable<MemberDeclarationSyntax> OptionalMultiReference()
+    {
+        InvocationExpressionSyntax getter;
+        MethodDeclarationSyntax tryGet;
+        switch (config.UnresolvedReferenceHandling)
         {
-            MultipleReferenceField(),
-            MultipleReferenceProperty(Call(ReferenceTargets().ToString())),
-            MultipleReferenceTargetsMethod(),
-            TryGetMultiple(InvocationExpression(ReferenceTargets()))
-        }.Concat(
-            LinkAdder([
-                SafeNodesVariableReference(),
-                AssureNotNullCall(reference),
-                AssureNotNullMembersCall(),
-                AddMultipleReferenceEmitterVariable(MemberAccess(FeatureField(reference),
-                    IdentifierName("Count"))),
-                EmitterCollectOldDataCall(),
-                SimpleAddRangeCall(),
-                EmitterNotifyCall(),
-                ReturnStatement(This())
-            ])
-        ).Concat(
-            LinkInserter([
-                AssureInRangeCall(),
-                SafeNodesVariableReference(),
-                AssureNotNullCall(reference),
-                AssureNotNullMembersCall(),
-                AddMultipleReferenceEmitterVariable(IdentifierName("index")),
-                EmitterCollectOldDataCall(),
-                SimpleInsertRangeCall(),
-                EmitterNotifyCall(),
-                ReturnStatement(This())
-            ])
-        ).Concat(
-            LinkRemover([
-                SafeNodesVariable(),
-                AssureNotNullCall(reference),
-                AssureNotNullMembersCall(),
-                SimpleRemoveAllCall(),
-                ReturnStatement(This())
-            ])
-        );
+            case UnresolvedReferenceHandling.ReturnAsNull:
+                getter = ReferenceTargetNullableTargetsCall();
+                tryGet = TryGetMultiple(
+                    ReferenceTargetNullableTargetsCall(),
+                    NullableType(AsType(reference.GetFeatureType()))
+                );
+                break;
+
+            case UnresolvedReferenceHandling.Throw:
+                getter = ReferenceTargetNonNullTargetsCall();
+                tryGet = TryGetMultipleReferences();
+                break;
+
+            default:
+                throw new ArgumentOutOfRangeException(nameof(config.UnresolvedReferenceHandling),
+                    config.UnresolvedReferenceHandling.ToString());
+        }
+
+        return new List<MemberDeclarationSyntax> { MultipleReferenceField(), MultipleReferenceProperty(getter), tryGet }
+            .Concat(
+                LinkAdder([
+                    SafeNodesVariableReference(),
+                    AssureNotNullCall(reference),
+                    AssureNotNullMembersCall(),
+                    AddMultipleReferenceEmitterVariable(MemberAccess(FeatureField(reference),
+                        IdentifierName("Count"))),
+                    EmitterCollectOldDataCall(),
+                    SimpleAddRangeCall(),
+                    EmitterNotifyCall(),
+                    ReturnStatement(This())
+                ])
+            ).Concat(
+                LinkInserter([
+                    AssureInRangeCall(),
+                    SafeNodesVariableReference(),
+                    AssureNotNullCall(reference),
+                    AssureNotNullMembersCall(),
+                    AddMultipleReferenceEmitterVariable(IdentifierName("index")),
+                    EmitterCollectOldDataCall(),
+                    SimpleInsertRangeCall(),
+                    EmitterNotifyCall(),
+                    ReturnStatement(This())
+                ])
+            ).Concat(
+                LinkRemover([
+                    SafeNodesVariable(),
+                    AssureNotNullCall(reference),
+                    AssureNotNullMembersCall(),
+                    SimpleRemoveAllCall(),
+                    ReturnStatement(This())
+                ])
+            );
+    }
+
+    private MethodDeclarationSyntax TryGetMultipleReferences() =>
+        AbstractTryGetMultiple(AsType(reference.GetFeatureType()))
+            .WithExpressionBody(ArrowExpressionClause(
+                InvocationExpression(
+                        GenericName(Identifier("TryGetReference"))
+                            .WithTypeArgumentList(TypeArgumentList(SingletonSeparatedList(
+                                AsType(reference.GetFeatureType())
+                            )))
+                    )
+                    .WithArgumentList(ArgumentList(SeparatedList([
+                        Argument(FeatureField(reference)),
+                        Argument(IdentifierName(FeatureTryGetParam())).WithRefOrOutKeyword(Token(SyntaxKind.OutKeyword))
+                    ])))
+            ))
+            .WithSemicolonToken(Token(SyntaxKind.SemicolonToken));
 
     private ExpressionStatementSyntax AssureNotNullMembersCall() =>
         ExpressionStatement(Call("AssureNotNullMembers",
@@ -265,18 +385,44 @@ public class FeatureGeneratorReference(Classifier classifier, Reference referenc
                 Collection([]))
             .WithModifiers(AsModifiers(SyntaxKind.PrivateKeyword, SyntaxKind.ReadOnlyKeyword));
 
-    private PropertyDeclarationSyntax
-        MultipleReferenceProperty(InvocationExpressionSyntax getter) =>
-        PropertyDeclaration(AsType(typeof(IReadOnlyList<>), AsType(reference.Type)),
+    private PropertyDeclarationSyntax MultipleReferenceProperty(InvocationExpressionSyntax getter)
+    {
+        TypeSyntax returnType;
+        InvocationExpressionSyntax initer;
+
+        switch (config.UnresolvedReferenceHandling)
+        {
+            case UnresolvedReferenceHandling.ReturnAsNull:
+                returnType = NullableType(AsType(reference.Type));
+                initer = InvocationExpression(
+                    LinkAdd(reference),
+                    AsArguments([PostfixUnaryExpression(
+                        SyntaxKind.SuppressNullableWarningExpression,
+                        IdentifierName("value")
+                    )])
+                );
+                break;
+
+            case UnresolvedReferenceHandling.Throw:
+                returnType = AsType(reference.Type);
+                initer = InvocationExpression(
+                    LinkAdd(reference),
+                    AsArguments([IdentifierName("value")])
+                );
+                break;
+
+            default:
+                throw new ArgumentOutOfRangeException(nameof(config.UnresolvedReferenceHandling), config.UnresolvedReferenceHandling.ToString());
+        }
+
+        return PropertyDeclaration(AsType(typeof(IReadOnlyList<>), returnType),
                 Identifier(FeatureProperty(reference).ToString()))
             .WithAccessorList(AccessorList(List([
                 AccessorDeclaration(SyntaxKind.GetAccessorDeclaration)
                     .WithExpressionBody(ArrowExpressionClause(getter))
                     .WithSemicolonToken(Token(SyntaxKind.SemicolonToken)),
                 AccessorDeclaration(SyntaxKind.InitAccessorDeclaration)
-                    .WithExpressionBody(ArrowExpressionClause(InvocationExpression(
-                        LinkAdd(reference),
-                        AsArguments([IdentifierName("value")])))
+                    .WithExpressionBody(ArrowExpressionClause(initer)
                     )
                     .WithSemicolonToken(Token(SyntaxKind.SemicolonToken))
             ])))
@@ -287,28 +433,15 @@ public class FeatureGeneratorReference(Classifier classifier, Reference referenc
             ]))
             .WithModifiers(AsModifiers(SyntaxKind.PublicKeyword))
             .Xdoc(XdocDefault(reference));
+    }
 
-    private MethodDeclarationSyntax SingleReferenceTargetMethod() =>
-        Method(ReferenceTarget().ToString(),
-                NullableType(AsType(reference.Type)))
-            .WithModifiers(AsModifiers(SyntaxKind.PrivateKeyword))
-            .WithExpressionBody(ArrowExpressionClause(CallGeneric("ReferenceTargetNullableTarget",
-                AsType(reference.Type), FeatureField(reference))))
-            .WithSemicolonToken(Token(SyntaxKind.SemicolonToken));
+    private InvocationExpressionSyntax ReferenceTargetNonNullTargetsCall() =>
+        CallGeneric("ReferenceTargetNonNullTargets",
+            AsType(reference.GetFeatureType()), FeatureField(reference), MetaProperty(reference));
 
-    private ExpressionSyntax ReferenceTarget() =>
-        IdentifierName($"{reference.Name.ToFirstUpper()}Target");
-
-    private MethodDeclarationSyntax MultipleReferenceTargetsMethod() =>
-        Method(ReferenceTargets().ToString(),
-                AsType(typeof(IImmutableList<>), AsType(reference.Type)))
-            .WithModifiers(AsModifiers(SyntaxKind.PrivateKeyword))
-            .WithExpressionBody(ArrowExpressionClause(CallGeneric("ReferenceTargetNullableTargets",
-                AsType(reference.GetFeatureType()), FeatureField(reference))))
-            .WithSemicolonToken(Token(SyntaxKind.SemicolonToken));
-
-    private ExpressionSyntax ReferenceTargets() =>
-        IdentifierName($"{reference.Name.ToFirstUpper()}Targets");
+    private InvocationExpressionSyntax ReferenceTargetNullableTargetsCall() =>
+        CallGeneric("ReferenceTargetNullableTargets",
+            AsType(reference.GetFeatureType()), FeatureField(reference), MetaProperty(reference));
 
     private LocalDeclarationStatementSyntax SafeNodesVariableReference() =>
         SafeNodesVariable(